% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_plot.R
\name{beta_binomial_plot}
\alias{beta_binomial_plot}
\title{Visualize a Beta-Binomial model}
\usage{
beta_binomial_plot(alpha, beta, x = NULL, n = NULL, prior = TRUE,
  likelihood = TRUE, posterior = TRUE)
}
\arguments{
\item{x, n}{Observed number of successes & trials, respectively}

\item{a, b}{Shape parameters of the Beta prior (Beta(a,b))}
}
\description{
Visualize the prior and posterior pdfs for a Beta-Binomial model.
}
\examples{
library(ggplot2)
beta_binomial_plot(a=2, b=1)
beta_binomial_plot(a=2, b=1, x=5, n=10)
}
