% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_summary.R
\name{beta_binomial_summary}
\alias{beta_binomial_summary}
\title{Summarize features of a Beta-Binomial model}
\usage{
beta_binomial_summary(alpha, beta, x = NULL, n = NULL)
}
\arguments{
\item{alpha, beta}{Shape parameters of the Beta prior (Beta(alpha,beta))}

\item{x, n}{Observed number of successes & trials}
}
\description{
Summarize the parameters, mean, mode, & variance of the prior and posterior pdfs for a Beta-Binomial model.
}
\examples{
beta_binomial_summary(alpha = 1, beta = 1)
beta_binomial_summary(alpha = 1, beta = 1, x = 5, n = 10)
}
