% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_power.R
\name{matrix_power}
\alias{matrix_power}
\title{Matrix power calculation}
\usage{
matrix_power(mat, n)
}
\arguments{
\item{mat}{a square matrix}

\item{n}{the power}
}
\description{
Calculate the nth power of a square matrix.  When the given matrix is a Markov chain transition matrix, this is equivalent to the n-step transition matrix.
}
\examples{
x <- matrix(c(1:9), ncol=3)
matrix_power(mat=x, n=1)
matrix_power(mat=x, n=2)
}
