% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_gamma_plot.R
\name{poisson_gamma_plot}
\alias{poisson_gamma_plot}
\title{Visualize a Poisson-Gamma model}
\usage{
poisson_gamma_plot(s, r, sum_x = NULL, n = NULL, prior = TRUE,
  likelihood = TRUE, posterior = TRUE)
}
\arguments{
\item{s, r}{Shape & rate parameters of the Gamma prior (Gamma(s,r))}

\item{sum_x, n}{Observed count total & sample size, respectively}
}
\description{
Visualize the prior and posterior pdfs for a Poisson-Gamma model.
}
\examples{
library(ggplot2)
poisson_gamma_plot(s=2, r=2)
poisson_gamma_plot(s=2, r=2, sum_x=10, n=5)
}
