% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_gamma_summary.R
\name{poisson_gamma_summary}
\alias{poisson_gamma_summary}
\title{Summarize features of a Poisson-Gamma model}
\usage{
poisson_gamma_summary(s, r, sum_x = NULL, n = NULL)
}
\arguments{
\item{s, r}{Shape & rate parameters of the Gamma prior (Gamma(s,r))}

\item{sum_x, n}{Observed count total & sample size, respectively}
}
\description{
Summarize the parameters, mean, mode, & variance of the prior and posterior pdfs for a Poisson-Gamma model.
}
\examples{
poisson_gamma_summary(s=2, r=2)
poisson_gamma_summary(s=2, r=2, sum_x=10, n=5)
}
