% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_mean_plot.R
\name{running_mean_plot}
\alias{running_mean_plot}
\title{Plot a running mean}
\usage{
running_mean_plot(x, se = FALSE, fill = "gray", alpha = 0.9)
}
\arguments{
\item{x}{a vector}

\item{se}{logical; if TRUE, naive confidence bands are included}

\item{fill}{fill color of confidence bands}

\item{alpha}{transparency of confidence bands}
}
\description{
Visualize the running, i.e. cumulative, mean of a vector.
}
\examples{
library(ggplot2)
x <- rnorm(1000)
running_mean_plot(x)
running_mean_plot(x, se=TRUE)
}
